requirements.txt
pdfplumber 
openai 
python-dotenv
pymupdf
numpy 
scikit-learn